feat: update OrderItemRepository

+++ src/main/java/com/splendore/repositories/restaurant/OrderItemRepository.java
@@ -3,5 +3,5 @@ package com.splendore.repositories.restaurant;
 import com.splendore.domain.restaurant.order.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
 }
